refactor(branch-protection): implement re-usable schema validation
For the time being, validation is only implemented for the attribute
types the custom schema validator is used with.
Others might be added later.

Please note that we do not error if the value is `null` since this
is supposed to be covered by the "AlsoRequires" validator.
