+++ internal/provider/boolvalidator/requires_true_if_configured.go
@@ -1,0 +1,18 @@
+package boolvalidator
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+
+	"terraform-provider-forgejo/internal/schemavalidator"
+)
+
+// RequiresTrueIfConfigured checks that any Bool values in the paths described by the
+// path.Expression are true if the current attribute value is configured to 'true'.
+// If you require the value described the the path.Expression to be set,
+// combine this validator with the "AlsoRequires" validator.
+func RequiresTrueIfConfigured(expressions ...path.Expression) validator.Bool {
+	return &schemavalidator.RequiresTrueIfConfiguredValidator{
+		Expressions: expressions,
+	}
+}
